Join test_await_jsproxy into test_jsproxy
