+++ src/tests/test_jsproxy.py
@@ -189,3 +189,123 @@ def test_jsproxy_kwargs(selenium):
         )
         == 5
     )
+
+
+import time
+
+startup = """
+import asyncio
+class DumbLoop(asyncio.AbstractEventLoop):
+    def create_future(self):
+        fut = asyncio.Future(loop=self)
+        old_set_result = fut.set_result
+        old_set_exception = fut.set_exception
+        def set_result(a):
+            print("set_result:", a)
+            old_set_result(a)
+        fut.set_result = set_result
+        def set_exception(a):
+            print("set_exception:", a)
+            old_set_exception(a)
+        fut.set_exception = set_exception
+        return fut
+
+    def get_debug(self):
+        return False
+
+asyncio.set_event_loop(DumbLoop())
+"""
+
+
+def test_await_jsproxy(selenium):
+    selenium.run(startup)
+    selenium.run(
+        """
+        def prom(res,rej):
+            global resolve
+            resolve = res
+        from js import Promise
+        p = Promise.new(prom)
+        async def temp():
+            x = await p
+            return x + 7
+        resolve(10)
+        c = temp()
+        r = c.send(None)
+        """
+    )
+    time.sleep(0.01)
+    msg = "StopIteration: 17"
+    with pytest.raises(selenium.JavascriptException, match=msg):
+        selenium.run(
+            """
+            c.send(r.result())
+            """
+        )
+
+
+def test_await_fetch(selenium):
+    selenium.run(startup)
+    selenium.run(
+        """
+        from js import fetch, window
+        async def test():
+            response = await fetch("console.html")
+            result = await response.text()
+            print(result)
+            return result
+        fetch = fetch.bind(window)
+
+        c = test()
+        r1 = c.send(None)
+        """
+    )
+    time.sleep(0.1)
+    selenium.run(
+        """
+        r2 = c.send(r1.result())
+        """
+    )
+    time.sleep(0.1)
+    msg = "StopIteration: <!doctype html>"
+    with pytest.raises(selenium.JavascriptException, match=msg):
+        selenium.run(
+            """
+            c.send(r2.result())
+            """
+        )
+
+
+def test_await_error(selenium):
+    selenium.run_js(
+        """
+        async function async_js_raises(){
+            console.log("Hello there???");
+            throw new Error("This is an error message!");
+        }
+        window.async_js_raises = async_js_raises;
+        function js_raises(){
+            throw new Error("This is an error message!");
+        }
+        window.js_raises = js_raises;
+        """
+    )
+    selenium.run(startup)
+    selenium.run(
+        """
+        from js import async_js_raises, js_raises
+        async def test():
+            c = await async_js_raises()
+            return c
+        c = test()
+        r1 = c.send(None)
+        """
+    )
+    msg = "This is an error message!"
+    with pytest.raises(selenium.JavascriptException, match=msg):
+        # Wait for event loop to go around for chome
+        selenium.run(
+            """
+            r2 = c.send(r1.result())
+            """
+        )
